Splits common file and fixes import issues
