+++ models/experimental/panoptic_deeplab/reference/aspp.py
@@ -4,7 +4,7 @@
 import torch.nn as nn
 import torch
 from torch import Tensor
-from models.experimental.panoptic_deeplab.tt.common import Conv2d
+from models.experimental.panoptic_deeplab.reference.utils import Conv2d
 
 
 class ASPPModel(torch.nn.Module):
