--- models/experimental/panoptic_deeplab/reference/panoptic_deeplab_weights_download.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-# Download your pretrained model:
-wget https://dl.fbaipublicfiles.com/detectron2/PanopticDeepLab/Cityscapes-PanopticSegmentation/panoptic_deeplab_R_52_os16_mg124_poly_90k_bs32_crop_512_1024_dsconv/model_final_23d03a.pkl -O models/experimental/panoptic_deeplab/reference/Panoptic_Deeplab_R52.pkl
