+++ models/experimental/panoptic_deeplab/tests/test_resnet52_bottleneck.py
@@ -10,8 +10,9 @@ from ttnn.model_preprocessing import preprocess_model_parameters
 from tests.ttnn.utils_for_testing import check_with_pcc
 from models.experimental.panoptic_deeplab.tt.bottleneck import TTBottleneck, get_bottleneck_optimisation
 from models.experimental.panoptic_deeplab.tt.custom_preprocessing import create_custom_mesh_preprocessor
-from models.experimental.panoptic_deeplab.tt.common import load_torch_model_state, Conv2d
+from models.experimental.panoptic_deeplab.common import load_torch_model_state
 from models.experimental.panoptic_deeplab.reference.resnet52_bottleneck import Bottleneck
+from models.experimental.panoptic_deeplab.reference.utils import Conv2d
 
 
 class BottleneckTestInfra:
