+++ models/experimental/panoptic_deeplab/tt/head.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ttnn
-from models.experimental.panoptic_deeplab.tt.common import TTConv2D, TTUpsample
+from models.experimental.panoptic_deeplab.tt.utils import TTConv2D, TTUpsample
 from dataclasses import dataclass
